refactor: move regex101 endpoint to page-local route, remove debug logs
